fix: move chat types to own folder
