+++ packages/types/src/documents/global/index.ts
@@ -11,3 +11,4 @@ export * from "./apikeys"
 export * from "./devInfo"
 export * from "./agents"
 export * from "./ai"
+export * from "./chat"
